requirements.txt
boto3
qdrant-client
openai
streamlit
python-dotenv
pillow